# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for LimeTV project \"A06367\"",
        "title": "LimeTV API",
        "contact": {
            "name": "Lelya Kochkarova, Ivan Chernov, Slava Rianov",
            "email": "kochkarova.lelya@gmail.com, chernov-ivan.1998@yandex.ru, slavarianov@yandex.ru"
        },
        "version": "v0.1.0"
    },
    "host": "http://3.67.182.34:8080",
    "basePath": "/",
    "paths": {
        "/actors/actor{id}": {
            "get": {
                "description": "Get details of actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actors"
                ],
                "summary": "Get details of actor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Actors"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/films/film/{film_id}": {
            "get": {
                "description": "Get details of film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get details of film",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Film"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/films/genre/{genre}": {
            "get": {
                "description": "Get films of genre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get films of genre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Боевик",
                        "name": "genre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/films/selection/actor/{actor_id}": {
            "get": {
                "description": "Get details of actor from selection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get details of actor from selection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "actor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/films/selection/user/personal": {
            "get": {
                "description": "Get details of personal film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get details of personal film",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Film"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/films/selection/{selection}": {
            "get": {
                "description": "Get details of films",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get details of films",
                "parameters": [
                    {
                        "type": "string",
                        "description": "КАКАЯ-ТО СТРОКА",
                        "name": "selection",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "options": {
                "description": "Get details of films",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get details of films",
                "parameters": [
                    {
                        "type": "string",
                        "description": "КАКАЯ-ТО СТРОКА",
                        "name": "selection",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/auth": {
            "get": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "SSID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "options": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/profile/{id}": {
            "get": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Profile"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/profile/{id}/follow": {
            "post": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/profile/{id}/unfollow": {
            "delete": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "768eb570-2e0e-11ec-8d3d-0242ac130004",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "description": "Get details of all actors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get details of all actors",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "SSID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Actors": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "./cmd/local/y.png"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "1968-03-03T02:18:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Английский актёр, наиболее известный по роли Джеймса Бонда."
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "{'Достать ножи'",
                        " 'Казино рояль'",
                        " 'Девушка с татуировкой дракона'}"
                    ]
                },
                "height": {
                    "type": "number",
                    "example": 1.78
                },
                "id": {
                    "type": "string",
                    "example": "768eb570-2e0e-11ec-8d3d-0242ac130004"
                },
                "name": {
                    "type": "string",
                    "example": "Дэниел"
                },
                "surname": {
                    "type": "string",
                    "example": "Крейг"
                }
            }
        },
        "models.Film": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "{'768eb570-2e0e-11ec-8d3d-0242ac130004'",
                        " '9ebe8b02-30e2-11ec-8d3d-0242ac130003'",
                        " 'a62bdb60-30e2-11ec-8d3d-0242ac130003'}"
                    ]
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "{'Райан Джонсон'",
                        " 'Рэм Бергман'}"
                    ]
                },
                "director": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "{'Райан Джонсон'",
                        " 'Леопольд Хьюз'}"
                    ]
                },
                "duration": {
                    "type": "integer",
                    "example": 130
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "{'Боевик'",
                        " 'Драма'}"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "768eb570-2e0e-11ec-8d3d-0242ac130004"
                },
                "language": {
                    "type": "string",
                    "example": "RU"
                },
                "release": {
                    "type": "string",
                    "example": "2019-10-28"
                },
                "title": {
                    "type": "string",
                    "example": "Достать ножи"
                },
                "year": {
                    "type": "integer",
                    "example": 2010
                }
            }
        },
        "models.LoginUser": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "Kalim1248"
                },
                "password": {
                    "type": "string",
                    "example": "dd81d9f0a8c5e7904931b8c9ccbf429b"
                }
            }
        },
        "models.Profile": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "subscribers": {
                    "type": "integer"
                },
                "subscriptions": {
                    "type": "integer"
                }
            }
        }
    }
}